# contracts/openapi/vvoice.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WarmupResponse": {
        "properties": {
          "loaded": {
            "title": "Loaded",
            "type": "boolean"
          },
          "loaded_languages": {
            "items": {
              "type": "string"
            },
            "title": "Loaded Languages",
            "type": "array"
          }
        },
        "required": [
          "loaded",
          "loaded_languages"
        ],
        "title": "WarmupResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "VassilStudio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/diagnostics/bundle": {
      "get": {
        "operationId": "diagnostics_bundle_diagnostics_bundle_get",
        "parameters": [
          {
            "description": "Include coarse Python, operating-system, and architecture metadata.",
            "in": "query",
            "name": "include_host_metadata",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include coarse Python, operating-system, and architecture metadata.",
              "title": "Include Host Metadata",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostics Bundle",
        "tags": [
          "system"
        ]
      }
    },
...
    "/warmup/asr": {
      "post": {
        "operationId": "warmup_asr_warmup_asr_post",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarmupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Warmup Asr",
        "tags": [
          "system"
        ]
      }
    },
    "/warmup/tts": {
      "post": {
        "operationId": "warmup_tts_warmup_tts_post",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarmupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Warmup Tts",
        "tags": [
          "system"
        ]
      }
    }
  }
}